+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Changed
+
+- Minor documentation tweaks ([#102][pr102])
+
+[pr102]: https://github.com/teloxide/teloxide-core/pull/102
+
 ## 0.3.0 - 2021-07-05
 
 ### Added