+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Minor documentation tweaks ([#102][pr102])
+- Remove `Self: 'static` buound on `RequesterExt::throttle` ([#102][pr102])
 
 [pr102]: https://github.com/teloxide/teloxide-core/pull/102