+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
-### Changed
+## 0.3.1 - 2021-07-07
 
 - Minor documentation tweaks ([#102][pr102])
 - Remove `Self: 'static` buound on `RequesterExt::throttle` ([#102][pr102])